Moved to the 'utils' package to be beside NamespacedName.
